+++ CHANGELOG.md
@@ -52,6 +52,10 @@ The following built-in themes are provided:
 - Add new theme plugin that provides components for switching between the dark and light themes
 - The toolbar Slot has new `SwitchTheme` and `SwitchThemeMenuItem`
 
+**Improvements**
+- Tweak toggle icons in bookmark items
+- The bookmark plugins provide new icons: `DownArrowIcon` and `RightArrowIcon`
+
 **Bug fixes**
 - Icons in menu items aren't centered horizontally